# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Accounting wrapper for the Lightning Network to support Taproot Assets and RGB.",
        "title": "LnHub",
        "contact": {
            "name": "BitTap",
            "url": "https://bittap.org"
        },
        "license": {
            "name": "GNU GPLv3",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "0.9.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v2/payments/bolt11": {
            "post": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Pay a bolt11 invoice",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Pay an invoice",
                "parameters": [
                    {
                        "description": "Invoice to pay",
                        "name": "PayInvoiceRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v2controllers.PayInvoiceRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v2controllers.PayInvoiceResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v2/payments/keysend": {
            "post": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Pay a node without an invoice using it's public key",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Make a keysend payment",
                "parameters": [
                    {
                        "description": "Invoice to pay",
                        "name": "KeySendRequestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v2controllers.KeySendRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v2controllers.KeySendResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v2/payments/keysend/multi": {
            "post": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Pay multiple nodes without an invoice using their public key",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Make multiple keysend payments",
                "parameters": [
                    {
                        "description": "Invoice to pay",
                        "name": "MultiKeySendRequestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v2controllers.MultiKeySendRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v2controllers.MultiKeySendResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "responses.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "v2controllers.KeySendRequestBody": {
            "type": "object",
            "required": [
                "amount",
                "asset_id",
                "destination"
            ],
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "asset_id": {
                    "type": "string"
                },
                "customRecords": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "custom_records": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "destination": {
                    "type": "string"
                },
                "memo": {
                    "type": "string"
                }
            }
        },
        "v2controllers.KeySendResponseBody": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "asset_id": {
                    "type": "integer"
                },
                "custom_records": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "description": {
                    "type": "string"
                },
                "description_hash": {
                    "type": "string"
                },
                "destination": {
                    "type": "string"
                },
                "fee": {
                    "type": "integer"
                },
                "payment_hash": {
                    "type": "string"
                },
                "payment_preimage": {
                    "type": "string"
                }
            }
        },
        "v2controllers.KeySendResult": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/responses.ErrorResponse"
                },
                "keysend": {
                    "$ref": "#/definitions/v2controllers.KeySendResponseBody"
                }
            }
        },
        "v2controllers.MultiKeySendRequestBody": {
            "type": "object",
            "properties": {
                "keysends": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v2controllers.KeySendRequestBody"
                    }
                }
            }
        },
        "v2controllers.MultiKeySendResponseBody": {
            "type": "object",
            "properties": {
                "keysends": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v2controllers.KeySendResult"
                    }
                }
            }
        },
        "v2controllers.PayInvoiceRequestBody": {
            "type": "object",
            "required": [
                "invoice"
            ],
            "properties": {
                "amount": {
                    "type": "integer",
                    "minimum": 0
                },
                "asset_id": {
                    "description": "NOTE: this is defaulting to bitcoin balance",
                    "type": "string",
                    "minLength": 1
                },
                "invoice": {
                    "type": "string"
                }
            }
        },
        "v2controllers.PayInvoiceResponseBody": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "asset_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "description_hash": {
                    "type": "string"
                },
                "destination": {
                    "type": "string"
                },
                "fee": {
                    "type": "integer"
                },
                "payment_hash": {
                    "type": "string"
                },
                "payment_preimage": {
                    "type": "string"
                },
                "payment_request": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/auth"
        }
    }
}